# docs/swagger.yaml
openapi: 3.0.0
info:
  title: PlatIAgro Projects API
  version: "0.2.0"
  description: >
    These are the docs for PlatIAgro Projects API.
    The endpoints below are usually accessed by the PlatIAgro Web-UI.
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: http://localhost:8080
...
paths:
...
  /projects/{projectId}/deployments:
    get:
      summary: "List all deployments."
      tags:
        - "Deployments"
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/Deployments"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      summary: "Create a new deployment."
      tags:
        - "Deployments"
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        $ref: "#/components/requestBodies/DeploymentPost"
      responses:
        "200":
          $ref: "#/components/responses/Deployments"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /projects/{projectId}/deployments/{deploymentId}:
    get:
      summary: "Detail a specific deployment by ID."
      tags:
        - "Deployments"
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: deploymentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/Deployment"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    patch:
      summary: "Update a deployment by ID."
      tags:
        - "Deployments"
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: deploymentId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        $ref: "#/components/requestBodies/DeploymentPatch"
      responses:
        "200":
          $ref: "#/components/responses/Deployment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      summary: "Delete a deployment by ID."
      tags:
        - "Deployments"
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
        - name: deploymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/Message"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /templates:
    get:
      summary: "List all templates sorted by name in natural sort order."
      tags:
        - "Templates"
      responses:
        "200":
          $ref: "#/components/responses/Templates"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      summary: "Create a new template from an experiment."
      tags:
        - "Templates"
      requestBody:
        $ref: "#/components/requestBodies/TemplatePost"
      responses:
        "200":
          $ref: "#/components/responses/Template"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /templates/{templateId}:
    get:
      summary: "Detail a specific template by ID."
      tags:
        - "Templates"
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/Template"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    patch:
      summary: "Update a template by ID."
      tags:
        - "Templates"
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        $ref: "#/components/requestBodies/TemplatePatch"
      responses:
        "200":
          $ref: "#/components/responses/Template"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      summary: "Delete a template by ID."
      tags:
        - "Templates"
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/Message"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
components:
  schemas:
...
    Deployment:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        experimentId:
          type: string
          format: uuid
        isActive:
          type: boolean
        name:
          type: string
        operators:
          type: array
          items:
            oneOf:
              - $ref: "#/components/schemas/Operator"
        position:
          type: integer
        status:
          type: string
        uuid:
          type: string
          format: uuid
    Deployments:
      type: array
      items:
        $ref: "#/components/schemas/Deployment"
...
    DeploymentExperiments:
      type: object
      properties:
        experiments:
          type: array
          items:
            type: string
            format: uuid
...
    DeploymentTemplate:
      type: object
      properties:
        templateId:
          type: string
          format: uuid
    Operator:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        dependencies:
          type: array
          items:
            type: string
            format: uuid
        parameters:
          type: object
          additionalProperties:
            type: string
        experimentId:
          type: string
          format: uuid
        positionX:
          type: number
          format: float
        positionY:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Template:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        tasks:
          type: array
          items:
            type: object
            properties:
              uuid:
                type: string
                format: uuid
              taskId:
                type: string
                format: uuid
              dependencies:
                type: array
                items:
                  type: string
                  format: uuid
              positionX:
                type: number
                format: float
              positionY:
                type: number
                format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Templates:
      type: array
      items:
        oneOf:
          - $ref: "#/components/schemas/Template"
  requestBodies:
...
    DeploymentPost:
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/DeploymentExperiments"
              - $ref: "#/components/schemas/DeploymentTemplate"
          examples:
            experiments:
              summary: From a list of experiments
              value:
                experiments: ["3fa85f64-5717-4562-b3fc-2c963f66afa6"]
            template:
              summary: From a template
              value:
                templateId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    DeploymentPatch:
      content:
        application/json:
          schema:
            type: object
            properties:
              isActive:
                type: boolean
              name:
                type: string
              position:
                type: integer
...
    TemplatePost:
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
              experimentId:
                type: string
                format: uuid
    TemplatePatch:
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
  responses:
...
    Deployment:
      description: ""
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Deployments"
    Deployments:
      description: ""
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Deployments"
...
    Template:
      description: ""
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Template"
    Templates:
      description: ""
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Templates"
    Message:
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    BadRequest:
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    NotFound:
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "The specified ... does not exist"
            required:
              - message
    InternalServerError:
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "An internal failure occurred."
            required:
              - message
    ServiceUnavailable:
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "The service is unavailable. Try your call again."
            required:
              - message